refactoring and features
- introduces abstract Minifier class
- removes file extension-check (this is done by the src filter already)
- adds a default name option for joined file (random names aren't helpful)
